Add the @autofill java annotation
the @autofill annotation causes the annotated property to be marked as autofillable when serialized into a json schema.

In addition to each @autofill property having a "type" property in the json schema, they will also have an "autofill" property which is always set to true.
If the property is not annotated with @autofill, it will not have the property "autofill" at all